Refactor sdk-react directory structure
* Split components/ into ui/ and providers/ to house respective types of components
* Update tsconfig
* Update package.json to include appropriate meta data
