+++ packages/sdk-react/package.json
@@ -1,14 +1,12 @@
 {
   "name": "@fusionauth/react-sdk",
-  "private": false,
   "version": "1.0.6",
-  "type": "module",
+  "private": false,
   "description": "FusionAuth solves the problem of building essential security without adding risk or distracting from your primary application",
+  "type": "module",
   "scripts": {
-    "dev": "vite",
     "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
   },
   "repository": {
     "type": "git",
@@ -24,31 +22,34 @@
     "npm": ">=10.0.0"
   },
   "homepage": "https://github.com/FusionAuth/fusionauth-react-sdk#readme",
-  "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
   "main": "dist/index.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist/**/*"
   ],
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
   "dependencies": {
     "classnames": "^2.3.2",
     "js-cookie": "^3.0.1"
   },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
   "devDependencies": {
     "@types/react": "^18.2.56",
     "@types/react-dom": "^18.2.19",
+    "@types/js-cookie": "3.0.6",
     "@typescript-eslint/eslint-plugin": "^7.0.2",
     "@typescript-eslint/parser": "^7.0.2",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "eslint": "^8.56.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "sass": "^1.56.1",
     "typescript": "^5.2.2",
-    "vite": "^5.1.4"
+    "vite": "^5.1.4",
+    "vite-plugin-dts": "^3.7.3"
   },
   "lint-staged": {
     "*.{ts,tsx}": "eslint --cache --fix",
