+++ packages/sdk-react/src/components/ui/FusionAuthLoginButton.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { useFusionAuth } from '../providers/FusionAuthProvider';
-import styles from '../styles/button.module.scss';
+import styles from '../../styles/button.module.scss';
 import classNames from 'classnames';
 
 interface Props {
