+++ packages/sdk-react/src/components/ui/FusionAuthRegisterButton.tsx
@@ -1,23 +1,28 @@
 import { FC } from 'react';
 import { useFusionAuth } from '../providers/FusionAuthProvider';
-import styles from '../styles/button.module.scss';
+import styles from '../../styles/button.module.scss';
 import classNames from 'classnames';
 
 interface Props {
+    state?: string;
     text?: string;
     className?: string;
 }
 
-export const FusionAuthLogoutButton: FC<Props> = ({ text, className }) => {
-    const { logout } = useFusionAuth();
+export const FusionAuthRegisterButton: FC<Props> = ({
+    state,
+    text,
+    className,
+}) => {
+    const { register } = useFusionAuth();
 
     return (
         <button
             className={classNames(styles.fusionAuthButton, className)}
             type="button"
-            onClick={() => logout()}
+            onClick={() => register(state ?? '')}
         >
-            {text ?? 'Logout'}
+            {text ?? 'Register Now'}
         </button>
     );
 };
