+++ packages/sdk-react/vite.config.ts
@@ -2,13 +2,8 @@
 // <reference types="vitest" />
 import { resolve } from 'node:path'
 import { defineConfig } from 'vite'
-// import dts from 'vite-plugin-dts'
+import dts from 'vite-plugin-dts'
 import react from '@vitejs/plugin-react-swc'
-
-// https://vitejs.dev/config/
-// export default defineConfig({
-//   plugins: [react()],
-// })
 
 
 
@@ -25,8 +20,8 @@ export default defineConfig({
     }
   },
   plugins: [
-    react()
-    // dts({ rollupTypes: true }),
+    react(),
+    dts({ rollupTypes: true }),
   ],
   // test: {},
 })
